First commit.
9 tabs, each with a different color.

Color itself set via a runtime attribute, names for the tabs set in the storyboard. (It takes three different places!)

Still to do: tables that allow push navigation, to test double-tapping a tab to go back up the navigation tree, and double-tapping a tab to scroll up to the top of the table.